Enable class constructors
The previous version was poluting the global namespace with the class
definitions and their members. This was preventing the use of constructors.
Now, only the class constructors are added to the global namespace, and
the class methods, which are only pertinent to instances, will only
be suggested when indexing from an instance of the given class.