Add link and type identity to tooltips
When an identifier has a known type, usually, that type
will be excplicitly stated in the autocomplete tooltip.
This helps understand what variables you are working with.

Furthermore, now, links to themoddingofisaac.com/docs are
provided with class and enumerator descriptions. I didn't
add links to function and attribute descriptions because
the most you can get out of it is already displayed in the
tooltip.